Refactor. Add second yearly subscription
