+++ app/src/main/java/io/ipoli/android/app/persistence/AndroidCalendarPersistenceService.java
@@ -12,7 +12,7 @@ import java.util.Map;
 import java.util.Set;
 
 import io.ipoli.android.app.events.AppErrorEvent;
-import io.ipoli.android.player.Player;
+import io.ipoli.android.player.data.Player;
 import io.ipoli.android.player.persistence.PlayerPersistenceService;
 import io.ipoli.android.quest.data.AndroidCalendarMapping;
 import io.ipoli.android.quest.data.Category;
