+++ app/src/main/java/io/ipoli/android/app/persistence/CalendarPersistenceService.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import io.ipoli.android.player.Player;
+import io.ipoli.android.player.data.Player;
 import io.ipoli.android.quest.data.AndroidCalendarMapping;
 import io.ipoli.android.quest.data.Category;
 import io.ipoli.android.quest.data.Quest;
