+++ app/src/main/java/io/ipoli/android/app/settings/SettingsActivity.java
@@ -70,7 +70,7 @@ import io.ipoli.android.app.ui.dialogs.TimePickerFragment;
 import io.ipoli.android.app.utils.DateUtils;
 import io.ipoli.android.app.utils.LocalStorage;
 import io.ipoli.android.app.utils.Time;
-import io.ipoli.android.player.Player;
+import io.ipoli.android.player.data.Player;
 import io.ipoli.android.player.UpgradeDialog;
 import io.ipoli.android.player.UpgradeManager;
 import io.ipoli.android.player.events.OpenAvatarStoreRequestEvent;
