+++ app/src/main/java/io/ipoli/android/app/sync/AndroidCalendarLoader.java
@@ -16,7 +16,7 @@ import java.util.Set;
 import io.ipoli.android.app.persistence.CalendarPersistenceService;
 import io.ipoli.android.app.utils.DateUtils;
 import io.ipoli.android.app.utils.LocalStorage;
-import io.ipoli.android.player.Player;
+import io.ipoli.android.player.data.Player;
 import io.ipoli.android.quest.data.Category;
 import io.ipoli.android.quest.data.Quest;
 import me.everything.providers.android.calendar.Event;
