+++ app/src/main/java/io/ipoli/android/feed/persistence/FirebaseFeedPersistenceService.java
@@ -17,7 +17,7 @@ import io.ipoli.android.app.persistence.FirebasePath;
 import io.ipoli.android.app.persistence.OnDataChangedListener;
 import io.ipoli.android.feed.data.Post;
 import io.ipoli.android.feed.data.Profile;
-import io.ipoli.android.player.Player;
+import io.ipoli.android.player.data.Player;
 
 /**
  * Created by Venelin Valkov <venelin@curiousily.com>
