+++ app/src/main/java/io/ipoli/android/player/persistence/CouchbasePlayerPersistenceService.java
@@ -14,7 +14,7 @@ import java.util.Map;
 import io.ipoli.android.app.persistence.BaseCouchbasePersistenceService;
 import io.ipoli.android.app.persistence.OnDataChangedListener;
 import io.ipoli.android.app.utils.StringUtils;
-import io.ipoli.android.player.Player;
+import io.ipoli.android.player.data.Player;
 
 /**
  * Created by Venelin Valkov <venelin@curiousily.com>
