+++ app/src/main/java/io/ipoli/android/player/persistence/PlayerPersistenceService.java
@@ -2,7 +2,7 @@ package io.ipoli.android.player.persistence;
 
 import io.ipoli.android.app.persistence.OnDataChangedListener;
 import io.ipoli.android.app.persistence.PersistenceService;
-import io.ipoli.android.player.Player;
+import io.ipoli.android.player.data.Player;
 
 /**
  * Created by Venelin Valkov <venelin@curiousily.com>
