+++ app/src/main/java/io/ipoli/android/store/fragments/AvatarStoreFragment.java
@@ -30,8 +30,8 @@ import io.ipoli.android.app.BaseFragment;
 import io.ipoli.android.app.events.EventSource;
 import io.ipoli.android.app.events.ScreenShownEvent;
 import io.ipoli.android.app.utils.DateUtils;
-import io.ipoli.android.player.Avatar;
-import io.ipoli.android.player.Player;
+import io.ipoli.android.player.data.Avatar;
+import io.ipoli.android.player.data.Player;
 import io.ipoli.android.player.events.AvatarBoughtEvent;
 import io.ipoli.android.player.persistence.PlayerPersistenceService;
 import io.ipoli.android.store.activities.StoreActivity;
