+++ app/src/main/java/io/ipoli/android/store/fragments/PetStoreFragment.java
@@ -32,8 +32,8 @@ import io.ipoli.android.app.events.EventSource;
 import io.ipoli.android.app.events.ScreenShownEvent;
 import io.ipoli.android.app.utils.DateUtils;
 import io.ipoli.android.app.utils.ResourceUtils;
-import io.ipoli.android.player.PetAvatar;
-import io.ipoli.android.player.Player;
+import io.ipoli.android.player.data.PetAvatar;
+import io.ipoli.android.player.data.Player;
 import io.ipoli.android.player.persistence.PlayerPersistenceService;
 import io.ipoli.android.store.activities.StoreActivity;
 import io.ipoli.android.store.adapters.PetStoreAdapter;
