+++ app/src/main/java/io/ipoli/android/store/fragments/UpgradeStoreFragment.java
@@ -42,7 +42,7 @@ import io.ipoli.android.app.ui.dialogs.AndroidCalendarsPickerFragment;
 import io.ipoli.android.app.ui.dialogs.LoadingDialog;
 import io.ipoli.android.app.utils.DateUtils;
 import io.ipoli.android.app.utils.LocalStorage;
-import io.ipoli.android.player.Player;
+import io.ipoli.android.player.data.Player;
 import io.ipoli.android.player.UpgradeManager;
 import io.ipoli.android.quest.data.Category;
 import io.ipoli.android.store.Upgrade;
